use server to serve the image so that the url is permanent

+++ mocacares_backend/event_platform/views/views_user.py
@@ -114,7 +114,8 @@ def set_user_info(request):
         user.username = request.POST['statement']
     if 'img' in request.POST:
         if user.portrait is not None:
-            user.portrait.delete()
+            if request.POST['img'] != user.portrait.image_url:
+                user.portrait.delete()
         user.portrait = UploadedImage.objects.get(image_url=request.POST['img'])
     if 'age' in request.POST:
         try:
